+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+- Added flag to ignore namespaces in check-kube-pods-pending
 
 ## 0.0.1 - 2016-03-03
 ### Added