+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ## [Unreleased]
 - Added flag to ignore namespaces in check-kube-pods-pending
+- check-kube-pods-service-available.rb: Will not mark a service is failed if any needed pod is running and ready
 
 ## 0.0.1 - 2016-03-03
 ### Added