+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - Added flag to ignore namespaces in check-kube-pods-pending
 - check-kube-service-available.rb: Will not mark a service is failed if any needed pod is running and ready
 - check-kube-service-available.rb: Added options to allow of pod pending for given time to be counted as valid
+- check-kube-service-available.rb: Fixed scope issue in main block that would cause a nil error
 
 ## 0.0.1 - 2016-03-03
 ### Added