+++ CHANGELOG.md
@@ -4,13 +4,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
 - Added flag to ignore namespaces in check-kube-pods-pending
 - check-kube-service-available.rb: Will not mark a service is failed if any needed pod is running and ready
 - check-kube-service-available.rb: Added options to allow of pod pending for given time to be counted as valid
-- check-kube-service-available.rb: Fixed scope issue in main block that would cause a nil error
 - Factored all checks to share a common base class for connecting to Kubernetes
 - Added flags to specify certificate authority and Kubernetes bearer token
 - Added flags to specify client certificate/key and in-cluster support
+
+### Fixed
+- check-kube-service-available.rb: Fixed scope issue in main block that would cause a nil error
+
+### Changed
+- Update to Rubocop 0.40 and cleanup
 
 ## 0.0.1 - 2016-03-03
 ### Added