+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+
+## [0.1.0] - 2016-05-15
 ### Added
 - Added flag to ignore namespaces in check-kube-pods-pending
 - check-kube-service-available.rb: Will not mark a service is failed if any needed pod is running and ready
@@ -17,6 +19,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ### Changed
 - Update to Rubocop 0.40 and cleanup
+- Update to kubeclient 1.1.3
 
 ## 0.0.1 - 2016-03-03
 ### Added